# swagger.yaml
openapi: 3.0.3
info:
  title: Pagar.me
  version: 0.2.4
  description: Backend challenge
  contact:
    email: juanfarberjob@gmail.com
servers:
  - url: "http://localhost:9999"
    description: Local API
tags:
  - name: General
    description: Test to see if the API is up
  - name: Transactions
    description: Transactions operations.

paths:
  /hello:
    get:
      tags:
        - General
      summary: Test to see if the API is up
      description: Test to see if the API is up
      responses:
        "200":
          description: Succesful operation.

  /tx/create:
    post:
      tags:
        - Transactions
      summary: Creates a transaction.
      description: Creates a transaction.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id_client:
                  description: Client ID.
                  type: number
                  example: 2
                value:
                  description: Tx value.
                  type: string
                  example: "10"
                payment_method:
                  description: Payment method.
                  type: string
                  example: "debit"
                card_number:
                  description: Card number.
                  type: string
                  example: "1122113311221133"
                card_holder:
                  description: Card holder.
                  type: string
                  example: "Juan Cito"
                expire_date:
                  description: Card expire date.
                  type: string
                  example: "2023-01-02"
                cvv:
                  description: Card security code.
                  type: string
                  example: "123"
                description:
                  description: Tx description.
                  type: string
                  example: "A very crazy yet no suspicious tx"
              required:
                - id_client
                - value
                - payment_method
                - card_number
                - card_holder
                - expire_date
                - cvv
      responses:
        "200":
          description: Succesful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "OK. Transaction created."
                  data:
                    description: Response data.
                    type: object
                    example:
                      {
                        "id_tx": 5,
                        "id_card": 1,
                        "card_type": "debit",
                        "card_number": "************1133",
                        "card_holder": "Juan Cito",
                        "id_client": 2,
                        "description": "A very crazy yet no suspicious tx",
                        "date": "2023-01-12 20:59:09.000000",
                        "value": "10.00",
                        "fee": "0.30",
                        "status": "P",
                      }
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "ERR. Invalid client id."
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "ERR. Binding parameters."

  /tx/list:
    get:
      tags:
        - Transactions
      summary: Creates a transaction.
      description: Creates a transaction.
      parameters:
        - in: query
          name: id_client
          description: Client ID
          schema:
            type: number
          example: 2
      responses:
        "200":
          description: Succesful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "OK. Transaction created."
                  data:
                    description: Response data.
                    type: object
                    example:
                      [
                        {
                          "id_tx": 7,
                          "card_type": "debit",
                          "card_number": "************1133",
                          "card_holder": "Juan Cito",
                          "id_client": 3,
                          "description": "A very crazy yet no suspicious tx",
                          "date": "2023-01-13 18:33:04.000000",
                          "value": "10.00",
                          "fee": "0.30",
                          "status": "P",
                        },
                        {
                          "id_tx": 8,
                          "card_type": "credit",
                          "card_number": "************5566",
                          "card_holder": "Juan Cito",
                          "id_client": 3,
                          "description": "A very crazy yet no suspicious tx 2",
                          "date": "2023-02-12 18:33:51.000000",
                          "value": "10.00",
                          "fee": "0.50",
                          "status": "W",
                        },
                      ]
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    description: Operation status message.
                    type: string
                    example: "ERR. Binding parameters."